+++ src/Runtime/Aggregator/MinAggregator.php
@@ -5,17 +5,28 @@ namespace Remorhaz\JSON\Path\Runtime\Aggregator;
 
 use Remorhaz\JSON\Data\Value\ScalarValueInterface;
 use Remorhaz\JSON\Data\Value\ValueInterface;
+use function array_search;
+use function min;
 
 final class MinAggregator extends UniqueNumericAggregator
 {
 
     protected function aggregateNumericData(array $dataList, ScalarValueInterface ...$elements): ?ValueInterface
     {
-        $elementIndex = array_search(min(...$dataList), $dataList, true);
-        if (false !== $elementIndex && isset($elements[$elementIndex])) {
+        $elementIndex = $this->findElementIndex($dataList);
+        if (isset($elementIndex, $elements[$elementIndex])) {
             return $elements[$elementIndex];
         }
 
-        throw new Exception\MinElementNotFoundException($dataList, $elements);
+        // @codeCoverageIgnoreStart
+        throw new Exception\MaxElementNotFoundException($dataList, $elements);
+        // @codeCoverageIgnoreEnd
+    }
+
+    private function findElementIndex(array $dataList): ?int
+    {
+        $elementIndex = array_search(min($dataList), $dataList, true);
+
+        return false === $elementIndex ? null : $elementIndex;
     }
 }