+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Issue #16: `min()`/`max()` aggregator functions fail on single-element array.
+
 ## [0.6.0] - 2019-11-06
 ### Changed
 - Some methods of `MapIndexInterface` renamed.